requirements.txt
streamlit==1.32.0
transformers==4.36.2
datasets==2.16.1
accelerate==0.25.0
bitsandbytes==0.41.1
torch==2.1.2
peft==0.7.1
trl==0.7.4
langchain==0.1.0
langchain-community==0.0.13
chromadb==0.4.22
sentence-transformers==2.2.2
python-dotenv==1.0.0
